refactor: member package 분리
