+++ backend/src/main/java/com/festago/auth/dto/LoginMemberDto.java
@@ -1,6 +1,6 @@
 package com.festago.auth.dto;
 
-import com.festago.domain.Member;
+import com.festago.zmember.domain.Member;
 
 public record LoginMemberDto(
     boolean isNew,
