+++ backend/src/main/java/com/festago/domain/MemberTicketRepository.java
@@ -1,6 +1,7 @@
 package com.festago.domain;
 
 import com.festago.stage.domain.Stage;
+import com.festago.zmember.domain.Member;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
