+++ backend/src/main/java/com/festago/zmember/application/MemberService.java
@@ -1,10 +1,10 @@
-package com.festago.application;
+package com.festago.zmember.application;
 
-import com.festago.domain.Member;
-import com.festago.domain.MemberRepository;
-import com.festago.dto.MemberProfileResponse;
 import com.festago.exception.ErrorCode;
 import com.festago.exception.NotFoundException;
+import com.festago.zmember.domain.Member;
+import com.festago.zmember.dto.MemberProfileResponse;
+import com.festago.zmember.repository.MemberRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
