+++ backend/src/main/java/com/festago/zmember/repository/MemberRepository.java
@@ -1,6 +1,7 @@
-package com.festago.domain;
+package com.festago.zmember.repository;
 
 import com.festago.auth.domain.SocialType;
+import com.festago.zmember.domain.Member;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
