+++ backend/src/test/java/com/festago/application/EntryServiceTest.java
@@ -10,7 +10,6 @@ import static org.mockito.BDDMockito.given;
 import com.festago.domain.EntryCode;
 import com.festago.domain.EntryCodePayload;
 import com.festago.domain.EntryState;
-import com.festago.domain.Member;
 import com.festago.domain.MemberTicket;
 import com.festago.domain.MemberTicketRepository;
 import com.festago.dto.EntryCodeResponse;
@@ -25,6 +24,7 @@ import com.festago.support.MemberTicketFixture;
 import com.festago.support.StageFixture;
 import com.festago.support.TimeInstantProvider;
 import com.festago.zfestival.domain.Festival;
+import com.festago.zmember.domain.Member;
 import java.time.Clock;
 import java.time.Instant;
 import java.time.LocalDateTime;
