+++ backend/src/test/java/com/festago/auth/application/AuthFacadeServiceTest.java
@@ -14,8 +14,8 @@ import com.festago.auth.dto.LoginMemberDto;
 import com.festago.auth.dto.LoginRequest;
 import com.festago.auth.dto.LoginResponse;
 import com.festago.auth.infrastructure.FestagoOAuth2Client;
-import com.festago.domain.Member;
 import com.festago.support.MemberFixture;
+import com.festago.zmember.domain.Member;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
