+++ backend/src/test/java/com/festago/auth/application/AuthServiceTest.java
@@ -10,10 +10,10 @@ import static org.mockito.BDDMockito.given;
 import com.festago.auth.domain.SocialType;
 import com.festago.auth.domain.UserInfo;
 import com.festago.auth.dto.LoginMemberDto;
-import com.festago.domain.Member;
-import com.festago.domain.MemberRepository;
 import com.festago.exception.NotFoundException;
 import com.festago.support.MemberFixture;
+import com.festago.zmember.domain.Member;
+import com.festago.zmember.repository.MemberRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
