+++ backend/src/test/java/com/festago/auth/application/integration/AuthFacadeServiceIntegrationTest.java
@@ -7,9 +7,9 @@ import com.festago.application.integration.ApplicationIntegrationTest;
 import com.festago.auth.application.AuthFacadeService;
 import com.festago.auth.domain.SocialType;
 import com.festago.auth.dto.LoginRequest;
-import com.festago.domain.Member;
-import com.festago.domain.MemberRepository;
 import com.festago.support.MemberFixture;
+import com.festago.zmember.domain.Member;
+import com.festago.zmember.repository.MemberRepository;
 import jakarta.persistence.EntityManager;
 import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
