+++ backend/src/test/java/com/festago/domain/MemberTest.java
@@ -3,6 +3,7 @@ package com.festago.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.festago.support.MemberFixture;
+import com.festago.zmember.domain.Member;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
