+++ backend/src/test/java/com/festago/domain/TicketTest.java
@@ -12,6 +12,7 @@ import com.festago.support.StageFixture;
 import com.festago.support.TicketFixture;
 import com.festago.ticket.domain.Ticket;
 import com.festago.zfestival.domain.Festival;
+import com.festago.zmember.domain.Member;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
