+++ backend/src/test/java/com/festago/support/MemberFixture.java
@@ -1,7 +1,7 @@
 package com.festago.support;
 
 import com.festago.auth.domain.SocialType;
-import com.festago.domain.Member;
+import com.festago.zmember.domain.Member;
 
 public class MemberFixture {
